+++ src/comm/all_to_all_comm.h
@@ -46,7 +46,7 @@ struct all_to_all_buffer
 /// output:
 ///     outer_hash_buffer_size: size of data: size of data received after all to all comm
 ///     outer_hash_buffer: buffer that holds data after all to all comm
-void all_to_all_comm(vector_buffer* vectorized_send_buffer, int vectorized_send_buffer_size, int* send_counts, u64 *outer_hash_buffer_size, u64 **recv_buffer, MPI_Comm comm);
+void all_to_all_comm(vector_buffer* vectorized_send_buffer, int vectorized_send_buffer_size, int* send_counts, int *outer_hash_buffer_size, u64 **recv_buffer, MPI_Comm comm);
 
 void comm_compaction_all_to_all(all_to_all_buffer compute_buffer, int **recv_buffer_offset_size, u64 **recv_buffer, MPI_Comm comm);
